+++ src/Encamina.Enmarcha.SemanticKernel.Abstractions/MemoryStoreHandlerBase.cs
@@ -14,13 +14,14 @@ public abstract class MemoryStoreHandlerBase : IMemoryStoreHandler
     /// <summary>
     /// Initializes a new instance of the <see cref="MemoryStoreHandlerBase"/> class.
     /// </summary>
-    /// <param name="memoryStore">The <see cref="IMemoryStore"/> to handle.</param>
-#pragma warning disable SKEXP0003 // Type is for evaluation purposes only and is subject to change or removal in future updates. Suppress this diagnostic to proceed.
-    protected MemoryStoreHandlerBase(IMemoryStore memoryStore)
-#pragma warning restore SKEXP0003 // Type is for evaluation purposes only and is subject to change or removal in future updates. Suppress this diagnostic to proceed.
+    /// <param name="memoryManager">A valid instance of <see cref="IMemoryManager"/> that manages the memory store handled by this instance.</param>
+    protected MemoryStoreHandlerBase(IMemoryManager memoryManager)
     {
-        MemoryStore = memoryStore;
+        MemoryManager = memoryManager;
     }
+
+    /// <inheritdoc/>
+    public virtual IMemoryManager MemoryManager { get; }
 
     /// <inheritdoc/>
     public virtual string CollectionNamePostfix { get; init; }
@@ -32,11 +33,6 @@ public abstract class MemoryStoreHandlerBase : IMemoryStoreHandler
     /// Gets the current collection of memory stores.
     /// </summary>
     protected IDictionary<string, MemoryStoreCollection> MemoryStoreCollectionInfo { get; } = new ConcurrentDictionary<string, MemoryStoreCollection>();
-
-    /// <summary>
-    /// Gets the current <see cref="IMemoryStore"/> handled by this instance.
-    /// </summary>
-    protected IMemoryStore MemoryStore { get; }
 
     /// <inheritdoc/>
     public virtual async Task<string> GetCollectionNameAsync(string collectionId, CancellationToken cancellationToken)
@@ -55,9 +51,9 @@ public abstract class MemoryStoreHandlerBase : IMemoryStoreHandler
             LastAccessUtc = DateTime.UtcNow,
         };
 
-        if (!await MemoryStore.DoesCollectionExistAsync(collectionName, cancellationToken))
+        if (!await MemoryManager.MemoryStore.DoesCollectionExistAsync(collectionName, cancellationToken))
         {
-            await MemoryStore.CreateCollectionAsync(collectionName, cancellationToken);
+            await MemoryManager.MemoryStore.CreateCollectionAsync(collectionName, cancellationToken);
         }
 
         return MemoryStoreCollectionInfo[collectionId].CollectionName;